test(uvu): remove jest, replace all tests with uvu, include nyc runner
