+++ .nycrc
@@ -1,0 +1,15 @@
+{
+  "extends": [
+    "@istanbuljs/nyc-config-typescript"
+  ],
+  "all": true,
+  "include": "src/**/*.ts",
+  "exclude": [
+    "src/__tests__/*.ts"
+  ],
+  "reporter": [
+    "text",
+    "lcov"
+  ],
+  "skip-full": true
+}
