--- jest.config.json
@@ -1,18 +1,0 @@
-{
-  "preset": "ts-jest",
-  "testMatch": ["<rootDir>/tests/**/*.(spec|test).ts"],
-  "testPathIgnorePatterns": ["<rootDir>/node_modules/"],
-  "coverageDirectory": "coverage",
-  "coverageReporters": ["lcov", "text"],
-  "collectCoverageFrom": ["<rootDir>/src/**/*.ts"],
-  "moduleNameMapper": {
-    "^@lib/(.*)$": "<rootDir>/src/$1",
-    "^@tests/(.*)$": "<rootDir>/tests/$1"
-  },
-  "watchPathIgnorePatterns": [
-    "<rootDir>/node_modules/",
-    "<rootDir>/dist/",
-    "<rootDir>/.git/",
-    "<rootDir>/.vscode/"
-  ]
-}
