+++ src/__tests__/@helpers/index.ts
@@ -16,9 +16,9 @@ export function result<T>(isOk: boolean, value: T): ReducedResult<T> {
 }
 
 export const should = {
-  // expose(exposed: Record<string, unknown>, ...exposees: Array<string>): void {
-  //   exposees.forEach((exposee) => assert.equal(exposed).toHaveProperty(exposee))
-  // },
+  expose(exposed: Record<string, unknown>, ...exposees: Array<string>): void {
+    exposees.forEach((exposee) => assert.ok(exposed[exposee]))
+  },
 
   matchState<T, R>(received: Result<T>, expected: ReducedResult<R>): void {
     assert.equal(received.isOk, expected.isOk)
@@ -34,3 +34,51 @@ export const should = {
     }
   }
 }
+
+export function testFailure<P extends Parser<unknown>>(input: string, parser: P) {
+  const actual = run(parser, input)
+  const expected = result(false, !actual.isOk ? actual.expected : actual.value)
+
+  should.matchState(actual, expected)
+}
+
+export function testSuccess<T, P extends Parser<unknown>>(input: string, value: T, parser: P) {
+  const actual = run(parser, input)
+  const expected = result(true, value)
+
+  should.matchState(actual, expected)
+}
+
+export const expectedCombinators = [
+  'chainl',
+  'choice',
+  'error',
+  'sepBy',
+  'many',
+  'map',
+  'mapTo',
+  'optional',
+  'sequence',
+  'takeLeft',
+  'takeMid',
+  'takeRight',
+  'takeSides'
+] as const
+
+export const expectedParsers = [
+  'defer',
+  'eof',
+  'eol',
+  'float',
+  'int',
+  'uint',
+  'letter',
+  'letters',
+  'nothing',
+  'regexp',
+  'rest',
+  'run',
+  'string',
+  'ustring',
+  'whitespace'
+] as const
