+++ src/__tests__/utils/unicode.data.ts
